Delete mount from working directory to a directory in the container
After discussing it with @snairdesai, we decided this should not be the default behavior. A mount would add the local project files to the container, but would not be connected to GitHub automatically. A better practice, more robust to potential merge conflicts, would be to git clone it within the container.